yb_backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1
bcrypt==3.2.2
bitarray==2.5.1
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==37.0.2
cytoolz==0.11.2
dataclasses==0.8
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
frozenlist==1.2.0
hexbytes==0.2.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
jwt==1.3.1
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
numpy==1.19.5
pandas==1.1.5
parsimonious==0.8.1
passlib==1.7.4
pkg_resources==0.0.0
protobuf==3.19.4
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
pyupbit==0.2.30
requests==2.27.1
rlp==2.0.1
schedule==1.1.0
six==1.16.0
toolz==0.11.2
typing_extensions==4.1.1
urllib3==1.26.9
varint==1.0.2
web3==5.29.2
websockets==9.1
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0